[TS migration] Migrate 'withCurrentUserPersonalDetails.js' HOC
